Example Corrections
Corrected some examples, as well as the expected Type for coordinate values (I have used Float values with success. It may be that there is some type conversion going on behind the scenes and Floats are converted to Strings or vice versa. Should probably confirm with UMS product and update docs again accordingly.