Simplify Clickable Mobile Content to SMS
On Stackoverflow I found a comment to combine the number/message separator, without the need to distinguish between iOS or Android:
https://stackoverflow.com/a/38708066

I've tested the proposed solution to combine the separator. While the order of the characters is important, I found it working on both iOS and Android.
* Tested on Samsung S7 Edge / Android 8.0.0 / Chrome 67.0.3396.87
* Tested on iPhone 6s / iOS 11.4 / Safari

When you're able to confirm this solution, I think we can greatly simplify the mobile content configuration, especially without the need to add additional code to the brand's website.