Update campaigns-android.md
IT appears that somewhere along the way in the In-App SDK development cycle an idea to present URL, section, and language from the App to LiveEngage got lost or dropped. The sample code provided on the developers page gives the reader the impression that they can send this information to LiveEngage - but in fact LiveEngage cannot interpret it as it stands currently. I am not sure if the issue lies with the design or the implementation. 

Whichever way the platform is supposed to work, the sample code should match to the expectation with appropriate comments to inform the developer what they can pass to LiveEngage and how it will be used.