:memo: Corrected Documentation
In this PR the documentation was corrected to reflect what is shown in the UI (0-7) instead (0-6). Additional to avoid future issues also clarified that scheduling will happen based on UTC time. Which usually is irrelevant for something like x minutes/hours, but has more relevance when it comes to running to specific times.