Update detailed-api.md
I just ran an experiment this morning. The phone_number attribute in the Standard Claim Set is being saved as the customerInfo.imei attribute.
The personal.contacts[].phone attribute is being saved as PersonalInfo.contacts[].phone.
Please determine whether the documentation should be updated or whether the currently deployed code is incorrect.